Use generic get_shmem_structure RPC for workqueues
