+++ kernel/src/arch/x86_64/tlb.rs
@@ -81,11 +81,23 @@ lazy_static! {
 
             channels
         } else {
-            // Get location of the work queues from the controller, who will have already created them in shared memory
-            use crate::arch::rackscale::get_workqueues::rpc_get_workqueues;
+            use crate::memory::{paddr_to_kernel_vaddr, PAddr};
+            use crate::arch::rackscale::get_shmem_structure::{rpc_get_shmem_structure, ShmemStructure};
             use crate::arch::rackscale::CLIENT_STATE;
-            let mut client = CLIENT_STATE.rpc_client.lock();
-            rpc_get_workqueues(&mut **client).expect("Failed to get shared memory workqueues from the controller")
+
+            // Get location of the work queues from the controller, who will have already created them in shared memory
+            let mut log_ptrs = [0u64; 1];
+            {
+                let mut client = CLIENT_STATE.rpc_client.lock();
+                rpc_get_shmem_structure(&mut **client, ShmemStructure::WorkQueues, &mut log_ptrs).expect("Failed to get nr log from controller");
+            }
+            let queue_ptr = paddr_to_kernel_vaddr(PAddr::from(log_ptrs[0]));
+            let local_workqueue_arc = unsafe {
+                Arc::from_raw(
+                    queue_ptr.as_u64() as *const Vec<ArrayQueue<(Arc<Shootdown>, TlbFlushHandle)>>
+                )
+            };
+            local_workqueue_arc
         }
     };
 }
